chore: refractor the folder of all tests related to books
