+++ app/src/androidTest/java/com/android/bookswap/ui/books/add/addBookTest.kt
@@ -1,4 +1,4 @@
-package com.android.bookswap.ui.book
+package com.android.bookswap.ui.books.add
 
 import android.content.Context
 import android.widget.Toast
@@ -13,8 +13,6 @@ import com.android.bookswap.data.BookGenres
 import com.android.bookswap.data.BookLanguages
 import com.android.bookswap.data.DataBook
 import com.android.bookswap.data.repository.BooksRepository
-import com.android.bookswap.ui.books.add.AddToBookScreen
-import com.android.bookswap.ui.books.add.createDataBook
 import com.android.bookswap.ui.navigation.NavigationActions
 import io.mockk.every
 import io.mockk.mockk
